only run the tests if curses is enabled
